+++ src/system_exclusive/mod.rs
@@ -19,11 +19,9 @@ mod show_control;
 pub use show_control::*;
 mod tuning;
 pub use tuning::*;
-mod util;
 
 use alloc::vec::Vec;
 use alloc::format;
-use self::util::checksum;
 
 use super::general_midi::GeneralMidi;
 use super::parse_error::*;
