+++ src/system_exclusive/notation.rs
@@ -3,7 +3,6 @@ use alloc::vec;
 use alloc::format;
 use crate::parse_error::*;
 use crate::util::*;
-use crate::system_exclusive::util::*;
 
 /// Indicates that the next MIDI clock message is the first clock of a new measure. Which bar
 /// is optionally indicated by this message.
